pyproject.toml
[project]
name = "drive-bot"
version = "0.1.0"
description = "Telegram bot that files scanned PDFs into Google Drive with AI-powered folder and name suggestions."
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=22.0,<23.0",
    "google-genai>=1.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "pydantic-settings>=2.0.0",
]
